+++ colour/models/ipt.py
@@ -111,7 +111,7 @@ def XYZ_to_IPT(XYZ: Domain1) -> Range1:
     --------
     >>> XYZ = np.array([0.20654008, 0.12197225, 0.05136952])
     >>> XYZ_to_IPT(XYZ)  # doctest: +ELLIPSIS
-    array([ 0.3842619...,  0.3848730...,  0.1888683...])
+    array([0.3842619..., 0.3848730..., 0.1888683...])
     """
 
     return XYZ_to_Iab(
@@ -161,7 +161,7 @@ def IPT_to_XYZ(IPT: Domain1) -> Range1:
     --------
     >>> IPT = np.array([0.38426191, 0.38487306, 0.18886838])
     >>> IPT_to_XYZ(IPT)  # doctest: +ELLIPSIS
-    array([ 0.2065400...,  0.1219722...,  0.0513695...])
+    array([0.2065400..., 0.1219722..., 0.0513695...])
     """
 
     return Iab_to_XYZ(
@@ -208,7 +208,7 @@ def IPT_hue_angle(IPT: Domain1) -> Range360:
     --------
     >>> IPT = np.array([0.96907232, 1, 1.12179215])
     >>> IPT_hue_angle(IPT)  # doctest: +ELLIPSIS
-    48.2852074...
+    np.float64(48.2852074...)
     """
 
     _I, P, T = tsplit(to_domain_1(IPT))
