+++ colour/quality/tm3018.py
@@ -145,7 +145,7 @@ def colour_fidelity_index_ANSIIESTM3018(
     >>> from colour import SDS_ILLUMINANTS
     >>> sd = SDS_ILLUMINANTS["FL2"]
     >>> colour_fidelity_index_ANSIIESTM3018(sd)  # doctest: +ELLIPSIS
-    70.1208244...
+    np.float64(70.1208244...)
     """
 
     if not additional_data:
